Added "floatPlacement" parameter. Also fixed scoping in helper functions
The helper functions were looking for the user's call at the top of the
call stack, i.e., with sys.call(1). That worked for simple call stacks
but not, for example, when calling build_vignettes(). The better way by
far seems to be to use sys.call(-1), and that is what I've implemented
with this commit.