Now checking for pdflatex and required packages only once per session.
latexTablePDF() was checking for the existence of pdflatex and the
required LaTeX packages every time it was run. These checks are
expensive; they make the function slow. Now the checks create variables
in the calling environment (.pdflatex_found and .latex_packages_found)
when they are first called. In future uses of latexTablePDF(), the
checks abort if those variables are found.