Improved checking of classes of regression objects in objList
The command that I was using, "classVec <- sapply(objList, class)", will
sometimes return a matrix with one column for each element of objList.
This will happen when the objects in objList have multiple classes.
Sometimes, the relevant classes (lm, plm, or ivgreg) will be the first
entries in the column; sometimes, they won't be. But the code in the
previous commit assumed that the relevant entries were the first
entries. This commit fixes the problem.