Fixed formatting problems that arose when formatNumbers in FALSE.
Fixed problem with handling of NA when formatNumbers is FALSE. ("NA" was
showing up in latexTable() output.)

Fixed problem with column padding in .tex output when formatNumbers is
FALSE. The major change is on line 489, where max(nchar(x)) has been
changed to max(nchar(x), na.rm = TRUE). The change speaks to problems
calculating widths when NA (not "NA") appears in mat, which I think
happens only when formatNumbers is FALSE.

Also a slight addition to the documentation of the formatNumbers
argument.