Fix failure to export print.regTable(). Updated vignette. Closes #21.
I had created but not exported print.regTable(). It was used during
development because I was loading the package scripts with
devtools::load_all(). But it wasn't easily available to end users, and
it wasn't used in the vignette. This commmit fixes the problem.

This commit also adds @export tags for all S3 methods. The @export tags
don't do anything now, because NAMESPACE is not created by roxygen. But
when I convert the legacy documentation to roxygen, I will let roxygen
overwrite NAMESPACE, and then those @export tags will tell roxygen to
automatically export the relevant methods.